Step 9.15: Extract StorageBreakdownModal
